fixed video ratio

+++ Assets/Scripts/UploadVideo.cs
@@ -1,4 +1,5 @@
-﻿using System.IO;
+﻿using System.Collections;
+using System.IO;
 using UnityEngine;
 using UnityEngine.UI;
 using SFB;
@@ -7,7 +8,7 @@ using NativeGalleryNamespace;
 
 public class UploadVideo : MonoBehaviour
 {
-    [SerializeField] bool isPlayTest = true;
+    [SerializeField] private bool isPlayTest = true;
     [SerializeField] private string storedVideoName = "myVideo.mp4";
     [SerializeField] private Button uploadButton;
     [SerializeField] private VideoPlayer videoPlayer;
@@ -15,26 +16,49 @@ public class UploadVideo : MonoBehaviour
 
     private void Awake()
     {
+        rawImageDisplay.gameObject.SetActive(false); // Make the RawImage not visible at the start
         uploadButton.onClick.AddListener(OnButtonClick);
     }
 
     public void OnButtonClick()
     {
         if(isPlayTest)
+        {
+            StopAndClearVideo();
             PickVideoDesktop();
+        }
         else
+        {
+            StopAndClearVideo();
             PickVideoMobile();
+        }
+    }
+
+    private void StopAndClearVideo()
+    {
+        if(videoPlayer.isPlaying)
+        {
+            videoPlayer.Stop();
+        }
+
+        string oldVideoPath = Path.Combine(Application.persistentDataPath, storedVideoName);
+        if(File.Exists(oldVideoPath))
+        {
+            File.Delete(oldVideoPath);
+        }
+
+        videoPlayer.clip = null;
     }
 
     public void PickVideoDesktop()
     {
         string[] paths = StandaloneFileBrowser.OpenFilePanel("Open Video File", "", "*", false);
-        if (paths.Length > 0)
+        if(paths.Length > 0)
         {
             File.Copy(paths[0], Path.Combine(Application.persistentDataPath, storedVideoName), true);
             PlayerPrefs.SetString("capturedVideo", storedVideoName);
             PlayerPrefs.Save();
-            PlayVideo("file://" + Path.Combine(Application.persistentDataPath, storedVideoName));
+            StartCoroutine(DelayPlayVideo("file://" + Path.Combine(Application.persistentDataPath, storedVideoName)));
         }
     }
 
@@ -42,15 +66,21 @@ public class UploadVideo : MonoBehaviour
     {
         NativeGallery.Permission permission = NativeGallery.GetVideoFromGallery((path) =>
         {
-            if (path != null)
+            if(path != null)
             {
                 PlayerPrefs.SetString("capturedVideo", path);
                 PlayerPrefs.Save();
-                PlayVideo("file:///" + path);
+                StartCoroutine(DelayPlayVideo("file://" + path));
             }
         }, "Select a video");
 
         Debug.Log("Permission result: " + permission);
+    }
+
+    private IEnumerator DelayPlayVideo(string videoUrl)
+    {
+        yield return new WaitForSeconds(0.1f);
+        PlayVideo(videoUrl);
     }
 
     private void PlayVideo(string videoUrl)
@@ -60,15 +90,46 @@ public class UploadVideo : MonoBehaviour
         
         videoPlayer.url = videoUrl;
         videoPlayer.Play();
+
+        rawImageDisplay.gameObject.SetActive(true); // Make the RawImage visible when video is ready to play
     }
 
+    public void CleanUpVideo()
+    {
+        if(videoPlayer.isPlaying)
+        {
+            videoPlayer.Stop();
+        }
+
+        string oldVideoPath = Path.Combine(Application.persistentDataPath, storedVideoName);
+        if (File.Exists(oldVideoPath))
+        {
+            File.Delete(oldVideoPath);
+        }
+
+        if (PlayerPrefs.HasKey("capturedVideo"))
+        {
+            PlayerPrefs.DeleteKey("capturedVideo");
+            PlayerPrefs.Save();
+        }
+
+       
+        videoPlayer.clip = null;
+
+        rawImageDisplay.gameObject.SetActive(false); // Make the RawImage not visible when older video is removed
+    }
+    
     private void RemoveVideo()
     {
         string fileName = PlayerPrefs.GetString("capturedVideo");
+        if (File.Exists(fileName))
+        {
+            File.Delete(fileName);
+        }
         PlayerPrefs.DeleteKey("capturedVideo");
-        PlayerPrefs.Save();
+        videoPlayer.clip = null;
 
-        File.Delete(fileName);
+        rawImageDisplay.gameObject.SetActive(false); // Make the RawImage not visible when video file is removed
     }
 
     private void LoadVideo()
@@ -76,7 +137,11 @@ public class UploadVideo : MonoBehaviour
         string fileName = PlayerPrefs.GetString("capturedVideo");
         if (File.Exists(fileName))
         {
-            PlayVideo("file:///" + fileName);
+            PlayVideo("file://" + fileName);
+        }
+        else
+        {
+            rawImageDisplay.gameObject.SetActive(false); // Make the RawImage not visible if fileName does not exist
         }
     }
 }
